+++ libft.h
@@ -14,7 +14,6 @@
 # define LIBFT_H
 # include <stdlib.h>
 # include <unistd.h>
-# include <string.h>
 
 typedef struct s_list
 {
@@ -22,17 +21,17 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-void			*ft_memchr(const void *s, int c, unsigned int n);
-void			*ft_memcpy(void *dest, const void *str, unsigned int num);
-void			*ft_memmove(void *dest, const void *str, unsigned int num);
-void			*ft_memset(void *ptr, int value, unsigned int num);
+void			*ft_memchr(const void *s, int c, size_t n);
+void			*ft_memcpy(void *dest, const void *str, size_t num);
+void			*ft_memmove(void *dest, const void *str, size_t num);
+void			*ft_memset(void *ptr, int value, size_t num);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putstr_fd(char *s, int fd);
 void			ft_putendl_fd(char *s, int fd);
 void			ft_putnbr_fd(int n, int fd);
 void			ft_striteri(char *s, void (*f)(unsigned int, char *));
-void			*ft_bzero(void *ptr, unsigned int num);
-void			*ft_calloc(unsigned int nmemb, unsigned int size);
+void			*ft_bzero(void *ptr, size_t num);
+void			*ft_calloc(size_t nmemb, size_t size);
 void			ft_lstadd_front(t_list **lst, t_list *new);
 void			ft_lstadd_back(t_list **lst, t_list *new);
 void			ft_lstdelone(t_list *lst, void (*del)(void *));
@@ -54,14 +53,14 @@ int				ft_isalpha(int caracter);
 int				ft_isascii(int c);
 int				ft_isdigit(int digit);
 int				ft_isprint(int c);
-int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
+int				ft_memcmp(const void *s1, const void *s2, size_t n);
 int				ft_strlen(const char *text);
-int				ft_strncmp(const char *str1, const char *str2, unsigned int n);
+int				ft_strncmp(const char *str1, const char *str2, size_t n);
 int				ft_toupper(int c);
 int				ft_tolower(int c);
 int				ft_lstsize(t_list *lst);
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int size);
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size);
+unsigned int	ft_strlcat(char *dst, const char *src, size_t size);
+unsigned int	ft_strlcpy(char *dest, const char *src, size_t size);
 t_list			*ft_lstnew(void *content);
 t_list			*ft_lstlast(t_list *lst);
 t_list			*ft_lstmap(t_list *lst, void *(*f)(void *)\
